+++ packages/_private/angular-demo/src/AllComponentParts/Pagination.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+import { SynPaginationComponent } from '@synergy-design-system/angular';
+
+@Component({
+  selector: 'demo-pagination',
+  standalone: true,
+  imports: [
+    SynPaginationComponent,
+  ],
+  template: `
+    <syn-pagination total-items="100"></syn-pagination>
+  `,
+})
+export class Pagination {}