+++ packages/_private/angular-demo/src/demos-template/demos-template.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { SynTabGroupComponent, SynTabComponent, SynTabPanelComponent } from '@synergy-design-system/angular';
+import { SynDividerComponent, SynTabGroupComponent, SynTabComponent, SynTabPanelComponent } from '@synergy-design-system/angular';
 import { CommonModule, type NgComponentOutlet } from '@angular/common';
 import { type SynTabShowEvent } from '@synergy-design-system/components';
 import { ActivatedRoute } from '@angular/router';
@@ -9,6 +9,7 @@ import { ActivatedRoute } from '@angular/router';
   standalone: true,
   imports: [
     CommonModule,
+    SynDividerComponent,
     SynTabGroupComponent,
     SynTabComponent,
     SynTabPanelComponent,
@@ -19,7 +20,6 @@ export class DemosTemplateComponent {
   activeDemo = '';
 
   demos: Array<[string, typeof NgComponentOutlet]> = [];
-
 
   constructor(private route: ActivatedRoute) {
     const routeDemos = this.route.snapshot.data['demos'];