+++ packages/_private/react-demo/src/AllComponentParts/Pagination.tsx
@@ -1,0 +1,3 @@
+export const Pagination = () => (
+  <syn-pagination total-items="100"></syn-pagination>
+);