+++ packages/_private/vanilla-demo/src/main.ts
@@ -15,10 +15,10 @@ import {
 import '@synergy-design-system/fonts';
 import '@synergy-design-system/demo-utilities/styles/font.css';
 
-import '@synergy-design-system/tokens/themes/sick2025_dark.css';
-import '@synergy-design-system/tokens/themes/sick2025_light.css';
 import '@synergy-design-system/tokens/themes/sick2018_dark.css';
 import '@synergy-design-system/tokens/themes/sick2018_light.css';
+import '@synergy-design-system/tokens/themes/sick2025_dark.css';
+import '@synergy-design-system/tokens/themes/sick2025_light.css';
 import '@synergy-design-system/components/index.css';
 import '@synergy-design-system/styles';
 import './app.css';