+++ packages/_private/vue-demo/src/AllComponentParts/DemoPagination.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts">
+import { SynVuePagination } from '@synergy-design-system/vue';
+</script>
+
+<template>
+  <SynVuePagination :total-items="100"></SynVuePagination>
+</template>