+++ packages/components/src/events/syn-pagination-page-changed.ts
@@ -1,0 +1,10 @@
+export type SynPaginationPageChangedEvent = CustomEvent<{
+  currentPage: number;
+  previousPage: number;
+}>;
+
+declare global {
+  interface GlobalEventHandlersEventMap {
+    'syn-pagination-page-changed': SynPaginationPageChangedEvent;
+  }
+}