+++ packages/metadata/data/layers/full/component/component:syn-pagination/react/SynPaginationJSXElement.ts
@@ -1,0 +1,7 @@
+export type SynPaginationJSXElement = SynCustomElement<
+  SynPagination,
+  [
+    ['syn-pagination-page-changed', SynPaginationPageChangedEvent],
+    ['syn-pagination-page-size-changed', SynPaginationPageSizeChangedEvent],
+  ]
+>;